Added ability to read from a file instead of user prompt
To make things more automated I added the ability to see if a file called slave.hostname exists. If it exists then the script will use the content of the file instead of asking for user input. If file does not exist then it will ask the user and create a file. 
Also changed the way the script gets the hostnames. Now it uses the user given hostnames instead of asking the server for it.